+++ lib/src/models/json_style_scheme.dart
@@ -7,6 +7,16 @@ class JsonQuotation {
 
   /// right quote
   final String rightQuote;
+
+  /// Json Quotation
+  ///
+  /// [JsonQuotation.arrowQuote] arrow quotation `<some_key>`
+  ///
+  /// [JsonQuotation.doubleQuote] double quotation `"some_key"`
+  ///
+  /// [JsonQuotation.singleQuote] single quotation `'some_key'`
+  ///
+  /// [JsonQuotation.same] left and right quotation are same `<leftQuotation>some_key<rightQuotation>`
   const JsonQuotation({
     this.leftQuote = '',
     this.rightQuote = '',
@@ -16,6 +26,16 @@ class JsonQuotation {
   const JsonQuotation.same(String quote)
       : leftQuote = quote,
         rightQuote = quote;
+
+  /// arrow quotation
+  static JsonQuotation arrowQuote =
+      const JsonQuotation(leftQuote: '<', rightQuote: '>');
+
+  /// double quotation
+  static JsonQuotation doubleQuote = const JsonQuotation.same('"');
+
+  /// single quotation
+  static JsonQuotation singleQuote = const JsonQuotation.same('\'');
 
   @override
   bool operator ==(Object other) {
@@ -41,11 +61,14 @@ class JsonStyleScheme {
   /// add double quotation on keys or not. default set to false
   final JsonQuotation? quotation;
 
+  final Widget? arrow;
+
   /// Json color scheme
   const JsonStyleScheme({
     this.keysStyle = const TextStyle(),
     this.valuesStyle = const TextStyle(),
     this.quotation = const JsonQuotation(),
+    this.arrow,
   });
 
   /// copy another JsonStyleScheme
@@ -53,11 +76,13 @@ class JsonStyleScheme {
     TextStyle? keysStyle,
     TextStyle? valuesStyle,
     JsonQuotation? quotation,
+    Widget? arrow,
   }) {
     return JsonStyleScheme(
       keysStyle: keysStyle ?? this.keysStyle,
       valuesStyle: valuesStyle ?? this.valuesStyle,
       quotation: quotation ?? this.quotation,
+      arrow: arrow ?? this.arrow,
     );
   }
 
@@ -68,6 +93,7 @@ class JsonStyleScheme {
       keysStyle: scheme.keysStyle,
       valuesStyle: scheme.valuesStyle,
       quotation: scheme.quotation,
+      arrow: scheme.arrow,
     );
   }
 
@@ -78,9 +104,10 @@ class JsonStyleScheme {
     return other is JsonStyleScheme &&
         other.keysStyle == keysStyle &&
         other.valuesStyle == valuesStyle &&
-        other.quotation == quotation;
+        other.quotation == quotation &&
+        other.arrow == arrow;
   }
 
   @override
-  int get hashCode => hashValues(keysStyle, valuesStyle, quotation);
+  int get hashCode => hashValues(keysStyle, valuesStyle, quotation, arrow);
 }